requirements.txt
python-telegram-bot==13.0
urllib3==1.26.7
environs==5.2.1
requests==2.26.0